Finished version of SIT encryption + Python module
