+++ Task03/sit_final/sit.h
@@ -1,48 +1,6 @@
-#include <stdio.h>
+#pragma once
+
 #include <stdint.h>
-
-const char P_TABLE[16] = {
-	0x03, 0x0F, 0x0E, 0x00,
-	0x05, 0x04, 0x0B, 0x0C,
-	0x0D, 0x0A, 0x09, 0x06,
-	0x07, 0x08, 0x02, 0x01
-};
-
-const char Q_TABLE[16] = {
-	0x09, 0x0E, 0x05, 0x06,
-	0x0A, 0x02, 0x03, 0x0C,
-	0x0F, 0x00, 0x04, 0x0D,
-	0x07, 0x0B, 0x01, 0x08
-};
-
-// Not used for encryption
-const char K1_PERM[16] = {
-	0x03, 0x02, 0x01, 0x00,
-	0x04, 0x05, 0x06, 0x07,
-	0x0B, 0x0A, 0x09, 0x08,
-	0x0C, 0x0D, 0x0E, 0x0F
-};
-
-const char K2_PERM[16] = {
-	0x00, 0x04, 0x08, 0x0C,
-	0x0D, 0x09, 0x05, 0x01,
-	0x02, 0x06, 0x0A, 0x0E,
-	0x0F, 0x0B, 0x07, 0x03
-};
-
-const char K3_PERM[16] = {
-	0x00, 0x01, 0x02, 0x03,
-	0x07, 0x06, 0x05, 0x04,
-	0x08, 0x09, 0x0A, 0x0B,
-	0x0F, 0x0E, 0x0D, 0x0C
-};
-
-const char K4_PERM[16] = {
-	0x0C, 0x08, 0x04, 0x00,
-	0x01, 0x05, 0x09, 0x0D,
-	0x0E, 0x0A, 0x06, 0x02,
-	0x03, 0x07, 0x0B, 0x0F
-};
 
 #define U8_TO_U16BE(p)	(*(p) << 8) | (*(p+1))
 #define U16BE_TO_U8(v, p)	do {	\
@@ -74,7 +32,7 @@ const char K4_PERM[16] = {
 #define GEN_SUBST(t, f)	do {									\
 	uint16_t tmp = 0;											\
 																\
-	for(char i = 0; i < 16; i++) {								\
+	for(uint8_t i = 0; i < 16; i++) {								\
 		tmp |= LR_SHIFT((f & (1 << t[i])), ((15 - i) - t[i]));	\
 	}															\
 																\
